% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveZeit.R
\name{saveZeit}
\alias{saveZeit}
\title{Saving data}
\usage{
saveZeit(df, format = c("txt", "sps", "sas", "dta"), path = getwd(),
  name = "zeit_df")
}
\arguments{
\item{df}{\code{data.frame}. Data set created.}

\item{format}{character. Format, one of "txt", "sps", "sas" or "dta".}

\item{path}{character. Directory path.}

\item{name}{character. File name.}
}
\description{
Internal function to save data sets to disk in different file formats.
}
\references{
\url{http://www.statmethods.net/input/exportingdata.html}
}

