% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitClient.R
\name{zeitClient}
\alias{zeitClient}
\title{Client status and API usage}
\usage{
zeitClient(api = Sys.getenv("zeit_api_key"), print = TRUE)
}
\arguments{
\item{api}{character. The personal api code. To request an API key see: \url{http://developer.zeit.de/quickstart/} This parameter is by default set to the R Environment.}

\item{print}{logical. If \code{TRUE} (default), the client information is printed.}
}
\value{
a list of information about the client and API usage
}
\description{
\code{zeitClient} lets you get information about your API access status and usage.
}
\examples{
\dontrun{
zeitClient()
}
}
\author{
Christian Graul (\email{christian.graul@gmail.com})
}

