% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitFrequencies.R
\name{zeitFrequencies}
\alias{zeitFrequencies}
\title{Calculate a frequency table}
\usage{
zeitFrequencies(ls, sort = c("year", "month", "week", "day"),
  save = c("txt", "sps", "sas", "dta"))
}
\arguments{
\item{ls}{list. List, created by \code{\link{fromZeit}}, to calculate the frequency table.}

\item{sort}{character. Specifies how the results are sorted. Possible options are \code{"year"} (default), \code{"month"}, \code{"week"} or \code{"day"}.}

\item{save}{character. Specifies the file for automatic saving of the resulting data frame. Possible formats are \code{"txt"}, \code{"sps"}, \code{"sas"} or \code{"dta"}. No data is saved per default.}
}
\value{
data frame
}
\description{
\code{zeitFrequencies} calculates a frequency table, i.e. a data frame including a continous list of dates in choosen sequences and the related frequencies. Returned objects can directly be used with \code{\link{zeitPlot}}.
}
\examples{
\dontrun{
# get data
mrkl <- fromZeit(q = "angela merkel", limit = "all", 
dateBegin = "2002-01-01", dateEnd = "2007-12-31")

# sort by month without saving
zeitFrequencies(ls = mrkl, sort = "month")

# sort by day and saved as .sas
zeitFrequencies(ls = mrkl, sort = "day", save = "sas")
}
}
\author{
Jan Dix (\email{jan.dix@uni-konstanz.de}), Jana Blahak (\email{jana.blahak@uni-konstanz.de}), Christian Graul (\email{christian.graul@gmail.com})
}
\seealso{
\code{\link{zeitSetApiKey}} \code{\link{fromZeit}} \code{\link{zeitToDf}} \code{\link{zeitPlot}}
}

