% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitPlot.R
\name{zeitPlot}
\alias{zeitPlot}
\title{Plot results of API request}
\usage{
zeitPlot(df, yTitle = "frequencies", xTitle = "time", title = NULL,
  grey = FALSE, absolute = TRUE, lowessFactor = 0.2)
}
\arguments{
\item{df}{data frame. Data frame that is the basis for the graph.}

\item{yTitle}{character. Labels the y-axis.}

\item{xTitle}{character. Labels the x-axis.}

\item{title}{character. Title of the graph.}

\item{grey}{logical. Indicates if print grey or in colored.}

\item{absolute}{logical. Indicates if absolute or relative frequencies are plotted on the y-axis.}

\item{lowessFactor}{numeric. "This gives the proportion of points in the plot which influence the smooth at each value. Larger values give more smoothness." For more details see \code{\link[stats]{lowess}}}
}
\value{
Plots a graph based on given data frame.
}
\description{
Function to plot the results of the data frames by frequencies and time automatically. Attention: Only works with the frequency data frames created by the internal function \code{zeitSorting} of the \code{rzeit} package.
}
\details{
If the data frame includes less than 20 observations, no trend line is plotted. The dotted line indicates the mean of frequencies.
}
\examples{
\dontrun{
zeitPlot(dfFrequenciesWeek, yTitle = "Week", xTitle = "Frequencies of articles")
}
}
\author{
Jan Dix, \email{jan.dix@uni-konstanz.de} Jana Blahak, \email{jana.blahak@uni-konstanz.de}
}
\seealso{
\code{\link{zeitSetApiKey}} \code{\link{fromZeit}} \code{\link{zeitFrequencies}} \code{\link{zeitToDf}}
}

