% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitSave.R
\name{zeitSave}
\alias{zeitSave}
\title{Save data to file system}
\usage{
zeitSave(df, format = c("txt", "sps", "sas", "dta"), path = getwd(),
  name = "zeit_df")
}
\arguments{
\item{df}{data.frame. Data set to be saved.}

\item{format}{character. Format, one of \code{"txt"}, \code{"sps"}, \code{"sas"} or \code{"dta"}.}

\item{path}{character. Directory path.}

\item{name}{character. File name.}
}
\description{
Function to save data sets to disk in different file formats.
}
\author{
Jan Dix (\email{jan.dix@uni-konstanz.de}), Jana Blahak (\email{jana.blahak@uni-konstanz.de}), Christian Graul (\email{christian.graul@gmail.com})
}
\references{
\url{http://www.statmethods.net/input/exportingdata.html}
}

