% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitSetApiKey.R
\name{zeitSetApiKey}
\alias{zeitSetApiKey}
\title{Sets the API Key to the environment}
\usage{
zeitSetApiKey(apiKey, path = paste(normalizePath("~/"), ".Renviron", sep =
  "/"))
}
\arguments{
\item{apiKey}{character. The personal api code. To request an API key see: \url{http://developer.zeit.de/quickstart/}}

\item{path}{character. Path where the enviornment is stored. Default is the normalized path.}
}
\description{
Function to set you API Key to the R environment when starting using \code{rzeit} package. Attention: You should only execute this functions once.
}
\examples{
\dontrun{
zeitSetApiKey(apiKey = "this_is_your_api_key", path = "")
}
}
\author{
Jan Dix (\email{jan.dix@uni-konstanz.de}), Jana Blahak (\email{jana.blahak@uni-konstanz.de})
}
\seealso{
\code{\link{fromZeit}} \code{\link{zeitFrequencies}} \code{\link{zeitToDf}} \code{\link{zeitPlot}}
}

