% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzeit.R
\docType{package}
\name{rzeit}
\alias{rzeit}
\alias{rzeit-package}
\title{Client for the ZEIT ONLINE Content API}
\description{
Interface to gather newspaper articles from ZEIT ONLINE, based on a 
multilevel query. Including sorting algorithms and graphical output options.
A personal API key is required for usage.
}
\references{
\url{http://developer.zeit.de}
}
\seealso{
\code{\link{fromZeit}} to expose a search in the ZEIT online archive,
\code{\link{zeitFrequencies}} to get a frequency table of the search results,
\code{\link{zeitPlot}} to plot results of a search,
\code{\link{zeitSetApiKey}} to set the API Key to the environment,
\code{\link{zeitToDf}} to get a data frame with sorted articles
}

