% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitFrequencies.R
\name{zeitFrequencies}
\alias{zeitFrequencies}
\title{Returns a frequency table in choosen format}
\usage{
zeitFrequencies(ls, sort = c("years", "months", "weeks", "days"),
  save = c("txt", "sps", "sas", "dta"))
}
\arguments{
\item{ls}{list. List which will be converted into the data frame. Attention: This function only works with the returned element of \code{\link{fromZeit}}.}

\item{sort}{character. Specifies how the results are sorted. Possible options are \code{year} (default), \code{month}, \code{week} or \code{day}. Note: \code{week} and \code{day} may take some time for a long time period!}

\item{save}{character. Specifies the file for automatic saving of the resulting data frame. Possible formats are \code{txt}, \code{sps}, \code{sas} or \code{dta}. No data is saved per default.}
}
\value{
data frame
}
\description{
The returned data frame includes a continous list of dates in choosen sequences and the related frequencies. The number of either day/week/month/year, are comparable to the data frame, created by \code{\link{zeitToDf}}.
}
\examples{
\dontrun{
# sort by month without saving
zeitFrequencies(terms, "month")

# sort by day and saved as sas
zeitFrequencies(terms, "day", "sas")
}
}
\author{
Jan Dix, \email{jan.dix@uni-konstanz.de} Jana Blahak, \email{jana.blahak@uni-konstanz.de}
}
\seealso{
\code{\link{zeitSetApiKey}} \code{\link{fromZeit}} \code{\link{zeitToDf}} \code{\link{zeitPlot}}
}

