% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitSplitSearch.R
\name{zeitSplitSearch}
\alias{zeitSplitSearch}
\title{Split search request}
\usage{
zeitSplitSearch(base, url, begin, end, q)
}
\arguments{
\item{begin}{Begin date - Restricts responses to results with publication dates of the date specified or later. In the form YYYY-MM-DD}

\item{end}{End date - Restricts responses to results with publication dates of the date specified or earlier. In the form YYYY-MM-DD}

\item{x}{url as character}
}
\description{
Allows to split the search in order to cover larger timespans.
}
\details{
\code{zeitSplitSearch} limits how many article are returend for each month, so that the search can cover a large timespan, even if one month originally contained an extremly large number of articles.
}

