% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitToDf.R
\name{zeitToDf}
\alias{zeitToDf}
\title{Convert list to data frame with sorted articles}
\usage{
zeitToDf(ls, sort = c("uuid", "title", "subtitle", "supertitle",
  "release_date", "href", "uri", "snippet", "teaser_text", "teaser_title"),
  save = c("txt", "sps", "sas", "dta"))
}
\arguments{
\item{ls}{list. List, created by \code{\link{fromZeit}}, that will be converted into the data frame.}

\item{sort}{character. Specifies how the results are sorted. Possible options are available variables: \code{"uuid"}, \code{"title"}, \code{"subtitle"}, \code{"supertitle"}, \code{"release_date"}, \code{"href"}, \code{"uri"}, \code{"snippet"}, \code{"teaser_text"} or \code{"teaser_title"} (or a subset of these, if the \code{fields} parameter was set in \code{fromZeit}-function).}

\item{save}{character. Specifies the file for automatic saving of the resulting data frame. Possible formats are \code{txt}, \code{sps}, \code{sas} or \code{dta}. No data is saved per default.}
}
\value{
data frame
}
\description{
\code{zeitToDf} converts the list object created by \code{\link{fromZeit}} to a data frame.
}
\examples{
\dontrun{
# get data
mrkl <- fromZeit(q = "angela merkel", limit = "all", 
dateBegin = "2002-01-01", dateEnd = "2007-12-31")

# convert to data frame without sorting and saving
zeitToDf(ls = mrkl)

# convert and sort by release_date
zeitToDf(ls = mrkl, sort = "release_date")

# convert and save as .txt
zeitToDf(ls = mrkl, save = "txt")
}
}
\author{
Jan Dix (\email{jan.dix@uni-konstanz.de}), Jana Blahak (\email{jana.blahak@uni-konstanz.de}), Christian Graul (\email{christian.graul@gmail.com})
}
\seealso{
\code{\link{zeitSetApiKey}} \code{\link{fromZeit}}
}

