% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitToDf.R
\name{zeitToDf}
\alias{zeitToDf}
\title{Returns a data frame with sorted articles}
\usage{
zeitToDf(ls, sort = c("years", "months", "weeks", "days"), save = c("txt",
  "sps", "sas", "dta"))
}
\arguments{
\item{ls}{list. List which will be converted into the data frame. Attention: This function only works with the returned element of \code{\link{fromZeit}}.}

\item{sort}{character. Specifies how the results are sorted. Possible options are \code{year}, \code{month}, \code{week} or \code{day}. Note: \code{week} and \code{day} may take some time for a long time period!}

\item{save}{character. Specifies the file for automatic saving of the resulting data frame. Possible formats are \code{txt}, \code{sps}, \code{sas} or \code{dta}. No data is saved per default.}
}
\value{
data frame
}
\description{
The returned data frame includes the articles, the related links and the number of either day/week/month/year, comparable to the data frame, created by \code{\link{zeitFrequencies}}.
}
\examples{
\dontrun{
# sort by month without saving
zeitToDf(terms, "month")

# sort by year and save as .sps
zeitToDf(terms, "day", "sps")
}
}
\author{
Jan Dix, \email{jan.dix@uni-konstanz.de} Jana Blahak, \email{jana.blahak@uni-konstanz.de}
}
\seealso{
\code{\link{zeitSetApiKey}} \code{\link{fromZeit}} \code{\link{zeitFrequencies}} \code{\link{zeitPlot}}
}

